% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-string.R
\name{arima_string}
\alias{arima_string}
\title{Forecast arima.string}
\usage{
arima_string(object, padding = FALSE)
}
\value{
A string
}
\description{
Forecast arima.string
}
\author{
Author(s) of \code{forecast} package
}
\keyword{internal}
