% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate-plot-list.R
\name{calibrate_and_plot}
\alias{calibrate_and_plot}
\title{Time-series Forecasting Simulator}
\usage{
calibrate_and_plot(..., type = "testing", .splits_obj, .data)
}
\arguments{
\item{...}{The workflow(s) you want to add to the function.}

\item{.splits_obj}{The splits object.}

\item{.data}{The full data set.}

\item{.type}{Either the training(splits) or testing(splits) data.}
}
\value{
The original time series, the simulated values and a some plots
}
\description{
This function is a helper function. It will take in a set of pre-processors and then
calibrate and plot them.
}
\details{
This function expects to take in workflows
}
\examples{
suppressPackageStartupMessages(library(modeltime))
suppressPackageStartupMessages(library(healthyR.data))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(timetk))
suppressPackageStartupMessages(library(ggplot2))
suppressPackageStartupMessages(library(plotly))
suppressPackageStartupMessages(library(tidyquant))

data <- healthyR_data \%>\%
 filter(ip_op_flag == "I") \%>\%
   select(visit_end_date_time) \%>\%
   rename(date_col = visit_end_date_time) \%>\%
   summarise_by_time(
       .date_var = date_col
       , .by     = "month"
       , value   = n()
  ) \%>\%
   filter_by_time(
       .date_var     = date_col
       , .start_date = "2012"
       , .end_date   = "2019"
   )


}
\author{
Steven P. Sanderson II, MPH
}
