% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate-misc-functions.R
\name{chr_assign}
\alias{chr_assign}
\title{Misc for boilerplate}
\usage{
chr_assign(name, value, cr = TRUE)
}
\value{
No return value, called for side effects
}
\description{
Misc for boilerplate
}
\keyword{internal}
