% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{get_gg_layers}
\alias{get_gg_layers}
\title{Get Random Walk \code{ggplot2} layers}
\usage{
get_gg_layers(df)
}
\value{
A \code{ggplot2} layers object
}
\description{
Get layers to add to a \code{ggplot} graph from the \code{\link[=ts_random_walk]{ts_random_walk()}} function.
}
\details{
\itemize{
\item Set the intercept of the initial value from the random walk
\item Set the max and min of the cumulative sum of teh random walks
}
}
\examples{
library(ggplot2)

df <- ts_random_walk()

df \%>\%
  ggplot(
    mapping = aes(
      x = x
      , y = cum_y
      , color = factor(run)
      , group = factor(run)
   )
 ) +
 geom_line(alpha = 0.8) +
 get_gg_layers(df)

}
\author{
Steven P. Sanderson II, MPH
}
