% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate-misc-functions.R
\name{get_recipe_call}
\alias{get_recipe_call}
\title{Misc for boilerplate}
\usage{
get_recipe_call(.rec_call)
}
\value{
No return value, called for side effects
}
\description{
Misc for boilerplate
}
\keyword{internal}
