% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-event-forward.R
\name{internal_ts_forward_event_tbl}
\alias{internal_ts_forward_event_tbl}
\title{Event Analysis}
\usage{
internal_ts_forward_event_tbl(.data, .horizon)
}
\arguments{
\item{.data}{The date.frame/tibble that holds the data.}

\item{.horizon}{How far do you want to look back or ahead.}
}
\value{
A tibble.
}
\description{
This is a function that sits inside of the \code{ts_time_event_analysis_tbl()}. It
is only meant to be used there. This is an internal function.
}
\details{
This is a helper function for \code{ts_time_event_analysis_tbl()} only.
}
\seealso{
Other Utility: 
\code{\link{auto_stationarize}()},
\code{\link{calibrate_and_plot}()},
\code{\link{internal_ts_backward_event_tbl}()},
\code{\link{internal_ts_both_event_tbl}()},
\code{\link{model_extraction_helper}()},
\code{\link{ts_get_date_columns}()},
\code{\link{ts_info_tbl}()},
\code{\link{ts_is_date_class}()},
\code{\link{ts_lag_correlation}()},
\code{\link{ts_model_auto_tune}()},
\code{\link{ts_model_compare}()},
\code{\link{ts_model_rank_tbl}()},
\code{\link{ts_model_spec_tune_template}()},
\code{\link{ts_qq_plot}()},
\code{\link{ts_scedacity_scatter_plot}()},
\code{\link{ts_to_tbl}()},
\code{\link{util_difflog_ts}()},
\code{\link{util_doublediff_ts}()},
\code{\link{util_log_ts}()},
\code{\link{util_singlediff_ts}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Utility}
