% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-ts-acceleration.R, R/step-ts-velocity.R
\name{required_pkgs.step_ts_acceleration}
\alias{required_pkgs.step_ts_acceleration}
\alias{required_pkgs.step_ts_velocity}
\title{Requited Packages}
\usage{
required_pkgs.step_ts_acceleration(x, ...)

required_pkgs.step_ts_velocity(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector

A character vector
}
\description{
Requited Packages

Required Packages
}
\keyword{internal}
