% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-fft.R
\name{tidy_fft}
\alias{tidy_fft}
\title{Tidy Style FFT}
\usage{
tidy_fft(
  .data,
  .value_col,
  .date_col,
  .frequency = 12L,
  .harmonics = 1L,
  .upsampling = 10L,
  .multi_harmonic = FALSE
)
}
\arguments{
\item{.data}{The data.frame/tibble you will pass for analysis.}

\item{.value_col}{The column that holds the data to be analyzed.}

\item{.date_col}{The column that holds the date.}

\item{.frequency}{The frequency of the data, 12 = monthly for example.}

\item{.harmonics}{How many harmonic waves do you want to produce.}

\item{.upsampling}{The upsampling of the time series.}

\item{.multi_harmonic}{A boolean TRUE/FALSE value that will decided if a
plot will be produced with all the harmonic waves on it from .harmonics.}
}
\value{
A list object.
}
\description{
Perform an fft using \code{\link[=fft]{fft()}} and return a tidier style output list with plots.
}
\details{

}
\examples{

}
\author{
Steven P. Sanderson II, MPH
}
