% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time-helpers.R
\name{ts_get_date_columns}
\alias{ts_get_date_columns}
\title{Get date or datetime variables (column names)}
\usage{
ts_get_date_columns(.data)
}
\arguments{
\item{.data}{An object of class \code{data.frame}}
}
\value{
A vector containing the column names that are of date/date-like classes.
}
\description{
Get date or datetime variables (column names)
}
\details{
\code{ts_get_date_columns} returns the column names of date or datetime variables
in a data frame.
}
\examples{
ts_to_tbl(AirPassengers) \%>\%
  ts_get_date_columns()

}
\seealso{
Other Utility: 
\code{\link{auto_stationarize}()},
\code{\link{calibrate_and_plot}()},
\code{\link{internal_ts_backward_event_tbl}()},
\code{\link{internal_ts_both_event_tbl}()},
\code{\link{internal_ts_forward_event_tbl}()},
\code{\link{model_extraction_helper}()},
\code{\link{ts_info_tbl}()},
\code{\link{ts_is_date_class}()},
\code{\link{ts_lag_correlation}()},
\code{\link{ts_model_auto_tune}()},
\code{\link{ts_model_compare}()},
\code{\link{ts_model_rank_tbl}()},
\code{\link{ts_model_spec_tune_template}()},
\code{\link{ts_qq_plot}()},
\code{\link{ts_scedacity_scatter_plot}()},
\code{\link{ts_to_tbl}()},
\code{\link{util_doublediff_ts}()},
\code{\link{util_log_ts}()},
\code{\link{util_singlediff_ts}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Utility}
