% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time-helpers.R
\name{ts_is_date_class}
\alias{ts_is_date_class}
\title{Check if an object is a date class}
\usage{
ts_is_date_class(.x)
}
\arguments{
\item{.x}{A vector to check}
}
\value{
Logical (TRUE/FALSE)
}
\description{
Check if an object is a date class
}
\examples{

seq.Date(from = as.Date("2022-01-01"), by = "day", length.out = 10) \%>\%
ts_is_date_class()

letters \%>\% ts_is_date_class()


}
\seealso{
Other Utility: 
\code{\link{auto_stationarize}()},
\code{\link{calibrate_and_plot}()},
\code{\link{internal_ts_backward_event_tbl}()},
\code{\link{internal_ts_both_event_tbl}()},
\code{\link{internal_ts_forward_event_tbl}()},
\code{\link{model_extraction_helper}()},
\code{\link{ts_get_date_columns}()},
\code{\link{ts_info_tbl}()},
\code{\link{ts_lag_correlation}()},
\code{\link{ts_model_auto_tune}()},
\code{\link{ts_model_compare}()},
\code{\link{ts_model_rank_tbl}()},
\code{\link{ts_model_spec_tune_template}()},
\code{\link{ts_qq_plot}()},
\code{\link{ts_scedacity_scatter_plot}()},
\code{\link{ts_to_tbl}()},
\code{\link{util_log_ts}()},
\code{\link{util_singlediff_ts}()}
}
\concept{Utility}
