% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette-color-blind.R
\name{ts_scale_color_colorblind}
\alias{ts_scale_color_colorblind}
\title{Provide Colorblind Compliant Colors}
\format{
Color_Blind
}
\usage{
ts_scale_color_colorblind(..., theme = "ts")
}
\description{
Provide Colorblind Compliant Colors
}
