% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-to-tbl.R
\name{ts_to_tbl}
\alias{ts_to_tbl}
\title{Coerce a time-series object to a tibble}
\usage{
ts_to_tbl(.data)
}
\arguments{
\item{.data}{The time-series object you want transformed into a \code{tibble}}
}
\value{

}
\description{
This function takes in a time-series object and returns it in a
\code{tibble} format.
}
\details{
This function makes use of \code{\link[timetk:tk_tbl]{timetk::tk_tbl()}} under the hood to obtain
the initial \code{tibble} object. After the inital object is obtained a new column
called \code{date_col} is constructed from the \code{index} column using \code{lubridate} if
an index column is returned.
}
\examples{

ts_to_tbl(BJsales)
ts_to_tbl(AirPassengers)

}
\seealso{
Other Utility: 
\code{\link{calibrate_and_plot}()},
\code{\link{model_extraction_helper}()},
\code{\link{ts_info_tbl}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Utility}
