% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-ts-acceleration.R
\name{required_pkgs.step_ts_acceleration}
\alias{required_pkgs.step_ts_acceleration}
\title{Requited Packages}
\usage{
required_pkgs.step_ts_acceleration(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
Requited Packages
}
\keyword{internal}
