% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-fft.R
\name{tidy_fft}
\alias{tidy_fft}
\title{Tidy Style FFT}
\usage{
tidy_fft(
  .data,
  .value_col,
  .date_col,
  .frequency = 12L,
  .harmonics = 1L,
  .upsampling = 10L
)
}
\arguments{
\item{.data}{The data.frame/tibble you will pass for analysis.}

\item{.value_col}{The column that holds the data to be analyzed.}

\item{.date_col}{The column that holds the date.}

\item{.frequency}{The frequency of the data, 12 = monthly for example.}

\item{.harmonics}{How many harmonic waves do you want to produce.}

\item{.upsampling}{The upsampling of the time series.}
}
\value{
A list object.
}
\description{
Perform an fft using \code{\link[=fft]{fft()}} and return a tidier style output list with plots.
}
\details{

}
\examples{
library(healthyR.data)
library(tidyverse)
library(timetk)

dat <- healthyR_data\%>\%
    filter(ip_op_flag == 'I') \%>\%
    summarise_by_time(
        .date_var = visit_end_date_time,
        .by = "month",
        V1 = n()
    ) \%>\%
    filter_by_time(
        .date_var = visit_end_date_time,
        .start_date = "2015",
        .end_date = "2019"
   )

a <- tidy_fft(
  .data = dat,
  .value_col = V1,
  .date_col = visit_end_date_time,
  .harmonics = 3,
  .frequency = 12
)

a$plots$max_har_plt
a$plots$harmonic_plt

}
\author{
Steven P. Sanderson II, MPH
}
