% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette-color-blind.R
\name{ts_scale_color_colorblind}
\alias{ts_scale_color_colorblind}
\title{Provide Colorblind Compliant Colors}
\usage{
ts_scale_color_colorblind(..., theme = "ts")
}
\description{
Provide Colorblind Compliant Colors
}
\seealso{
Other Color_Blind: 
\code{\link{ts_scale_fill_colorblind}()}
}
\concept{Color_Blind}
